rows: move fields related code to fields.go
