+++ rows.go
@@ -9,119 +9,10 @@
 package mysql
 
 import (
-	"database/sql"
 	"database/sql/driver"
 	"io"
 	"reflect"
-	"time"
 )
-
-var (
-	scanTypeNil        = reflect.TypeOf(nil)
-	scanTypeNullInt    = reflect.TypeOf(sql.NullInt64{})
-	scanTypeUint8      = reflect.TypeOf(uint8(0))
-	scanTypeInt8       = reflect.TypeOf(int8(0))
-	scanTypeUint16     = reflect.TypeOf(uint16(0))
-	scanTypeInt16      = reflect.TypeOf(int16(0))
-	scanTypeUint32     = reflect.TypeOf(uint32(0))
-	scanTypeInt32      = reflect.TypeOf(int32(0))
-	scanTypeUint64     = reflect.TypeOf(uint64(0))
-	scanTypeInt64      = reflect.TypeOf(int64(0))
-	scanTypeNullFloat  = reflect.TypeOf(sql.NullFloat64{})
-	scanTypeFloat32    = reflect.TypeOf(float32(0))
-	scanTypeFloat64    = reflect.TypeOf(float64(0))
-	scanTypeNullString = reflect.TypeOf(sql.NullString{})
-	scanTypeString     = reflect.TypeOf("")
-	scanTypeBytes      = reflect.TypeOf([]byte{})
-	scanTypeRawBytes   = reflect.TypeOf(sql.RawBytes{})
-	scanTypeTime       = reflect.TypeOf(time.Time{})
-	scanTypeUnknown    = reflect.TypeOf(new(interface{}))
-)
-
-type mysqlField struct {
-	tableName string
-	name      string
-	flags     fieldFlag
-	fieldType byte
-	decimals  byte
-}
-
-func (mf *mysqlField) scanType() reflect.Type {
-	switch mf.fieldType {
-	case fieldTypeNULL:
-		return scanTypeNil
-
-	case fieldTypeTiny:
-		if mf.flags&flagNotNULL != 0 {
-			if mf.flags&flagUnsigned != 0 {
-				return scanTypeUint8
-			}
-			return scanTypeInt8
-		}
-		return scanTypeNullInt
-
-	case fieldTypeShort, fieldTypeYear:
-		if mf.flags&flagNotNULL != 0 {
-			if mf.flags&flagUnsigned != 0 {
-				return scanTypeUint16
-			}
-			return scanTypeInt16
-		}
-		return scanTypeNullInt
-
-	case fieldTypeInt24, fieldTypeLong:
-		if mf.flags&flagNotNULL != 0 {
-			if mf.flags&flagUnsigned != 0 {
-				return scanTypeUint32
-			}
-			return scanTypeInt32
-		}
-		return scanTypeNullInt
-
-	case fieldTypeLongLong:
-		if mf.flags&flagNotNULL != 0 {
-			if mf.flags&flagUnsigned != 0 {
-				return scanTypeUint64
-			}
-			return scanTypeInt64
-		}
-		return scanTypeNullInt
-
-	case fieldTypeFloat:
-		if mf.flags&flagNotNULL != 0 {
-			return scanTypeFloat32
-		}
-		return scanTypeNullFloat
-
-	case fieldTypeDouble:
-		if mf.flags&flagNotNULL != 0 {
-			return scanTypeFloat64
-		}
-		return scanTypeNullFloat
-
-	case fieldTypeDecimal, fieldTypeNewDecimal, fieldTypeVarChar,
-		fieldTypeBit, fieldTypeEnum, fieldTypeSet, fieldTypeTinyBLOB,
-		fieldTypeMediumBLOB, fieldTypeLongBLOB, fieldTypeBLOB,
-		fieldTypeVarString, fieldTypeString, fieldTypeGeometry, fieldTypeJSON,
-		fieldTypeTime:
-		if mf.flags&flagNotNULL != 0 {
-			// alternatively we could return []byte or even RawBytes
-			return scanTypeString
-		}
-		return scanTypeNullString
-
-	case
-		fieldTypeDate, fieldTypeNewDate,
-		fieldTypeTimestamp, fieldTypeDateTime:
-
-		// TODO: NULL
-		// TODO: respect rows.mc.parseTime
-		return scanTypeTime
-
-	default:
-		return scanTypeUnknown
-	}
-}
 
 type resultSet struct {
 	columns []mysqlField
